+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 0.0.2
 
+## Breaking change
+
+- Changed the default variable replacer from `BracketVariableReplacer` to `DelimiterVariableReplacer`
+
+## Feature
+
+- Added `DelimiterVariableReplacer` with `{{` as prefix and `}}` as suffix by default
+
 ## CI
 
 - Added github actions to create a tag when a new version is released to pub.dev